pyproject.toml
[project]
name = "pylib"
version = "0.1.0"
description = "Reusable Python3 classes and functions."
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "boto3>=1.42.35",
    "chromadb>=1.4.1",
    "ebooklib>=0.20",
    "flagembedding>=1.3.5",
    "flashrank>=0.2.10",
    "langchain-text-splitters>=1.1.0",
    "llama-index-core>=0.14.13",
    "llama-index-embeddings-ollama>=0.8.6",
    "llama-index-llms-bedrock>=0.4.2",
    "llama-index-llms-ollama>=0.9.1",
    "llama-index-packs-raptor>=0.4.1",
    "llama-index-vector-stores-chroma>=0.5.5",
    "matplotlib>=3.10.8",
    "pdfplumber>=0.11.9",
    "pymysql>=1.1.2",
    "pypdf[crypto]>=6.6.2",
    "python-docx>=1.2.0",
    "pyyaml>=6.0.3",
    "rdflib>=7.5.0",
    "requests>=2.32.5",
    "ruamel-yaml>=0.19.1",
    "striprtf>=0.0.29",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lib"]
